api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "OCR api"
authors = ["Chris Birster <christopher.birster@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.3"
easyocr = "^1.7.1"
pillow = "^10.3.0"
torch = "^2.3.0"
numpy = "^1.26.4"
opencv-python-headless = "^4.9.0.80"
flask-cors = "^4.0.1"
openai = "^1.30.1"
python-dotenv = "^1.0.1"
pytz = "^2024.1"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
